GEN_AI_PROJECT/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
gitpython==3.1.40
gitignore-parser==0.1.11
